Added better option menu generation

+++ UI/Options/Options.cs
@@ -12,9 +12,9 @@ public partial class Options
 	[OptionGroup("Gameplay & Accessibility")]
 	public OptionDifficulty Difficulty {get; set;} = OptionDifficulty.Easy;
 	public OptionColorblindMode ColorBlindMode {get; set;} = OptionColorblindMode.None;
-	public float SpeedModifier {get; set;} = 1.0f;
+	public double SpeedModifier {get; set;} = 1.0f;
 	[OptionDesc("Extra time slowdown for speed critical moments")]
-	public float SpeedCritModifier {get; set;} = 1.0f;
+	public double SpeedCritModifier {get; set;} = 1.0f;
 	public bool GodMode {get; set;} = false;
 
 	[OptionGroup("Display")]
@@ -22,14 +22,14 @@ public partial class Options
 	public OptionWindowDisplay WindowDisplay {get; set;} = OptionWindowDisplay.Windowed;
 	public int FrameRate {get; set;} = 60;
 	public bool ScreenShake {get; set;} = true;
-	public float Brightness {get; set;} = 1.0f;
+	public double Brightness {get; set;} = 1.0f;
 
 	[OptionGroup("Volume")]
-	[OptionSlider] public float MasterVolume {get; set;} = 0.0f;
-	[OptionSlider] public float MusicVolume {get; set;} = 1.0f;
-	[OptionSlider] public float UIVolume {get; set;} = 1.0f;
-	[OptionSlider] public float PlayerVolume {get; set;} = 1.0f;
-	[OptionSlider] public float AmbientVolume {get; set;} = 1.0f;
+	[OptionSlider] public double MasterVolume {get; set;} = 0.0f;
+	[OptionSlider] public double MusicVolume {get; set;} = 1.0f;
+	[OptionSlider] public double UIVolume {get; set;} = 1.0f;
+	[OptionSlider] public double PlayerVolume {get; set;} = 1.0f;
+	[OptionSlider] public double AmbientVolume {get; set;} = 1.0f;
 
 	[OptionGroup("Video")]
 	public OptionGraphicsQuality GraphicsQuality {get; set;} = OptionGraphicsQuality.Medium;
